requirements.txt
pytest==8.2.2
allure-pytest==2.13.5
selene>=2.0.b4
